Move command related classes into commands package
