+++ Wurst Client/src/tk/wurst_client/bot/commands/Command.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package tk.wurst_client.bot;
+package tk.wurst_client.bot.commands;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
